requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
